// reference-mcp-servers/time/usrrname_worldtime-mcp-server_6413339_typescript/types/types.d.ts
export type WorldTimeResponse = {
  utc_offset: string;
  timezone: string;
  day_of_week: number;
  day_of_year: number;
  datetime: string;
  utc_datetime: string;
  unixtime: number;
  raw_offset: number;
  week_number: number;
  dst: boolean;
  abbreviation: string;
  dst_offset: number;
  dst_from: string | null;
  dst_until: string | null;
  client_ip: string;
};
// ...
type Zone = {
  countryCode: string;
  countryName: string;
  zoneName: string;
  gmtOffset: number;
  timestamp: number;
}
  
export type TimezoneDbBaseResponse = {
  status: "OK" | "ERROR";
  message: string;
}

export type ListTimezonesResponse = TimezoneDbBaseResponse & {
  status: "OK" | "ERROR";
  message: string;
  zones: Zone[];
}

export type GetTimeZoneResponse = TimezoneDbBaseResponse & {
  "countryCode": string;
  "countryName": string;
  "regionName": string;
  "cityName": string;
  "zoneName": string;
}

export type TimezoneConversionResponse = TimezoneDbBaseResponse & {
  fromZoneName: string;
  fromAbbreviation: string;
  fromTimestamp: number;
  toZoneName: string;
  toAbbreviation: string;
  toTimestamp: number;
  offset: number;
}

export type TimezoneDbGetTimeZoneResponse = TimezoneDbBaseResponse & {
  "countryCode": string;
  "countryName": string;
  "regionName": string;
  "cityName": string;
  "zoneName": string;
  "abbreviation": string;
  "gmtOffset": number;
  "dst": "0"|"1";
  "zoneStart": number;
  "zoneEnd": number;
  "nextAbbreviation": string;
  "timestamp": number;
  "formatted": string;
}